Handle two changes related to FileFormatWriter in Spark 340
