+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/GpuV1WriteUtils.scala
@@ -1,0 +1,69 @@
+/*
+ * Copyright (c) 2022, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark.sql.rapids
+
+import ai.rapids.cudf.ColumnVector
+import com.nvidia.spark.rapids._
+
+import org.apache.spark.internal.Logging
+import org.apache.spark.sql.catalyst.expressions.{Attribute, AttributeSet, Expression, NamedExpression}
+import org.apache.spark.sql.types.{DataType, StringType}
+
+object GpuV1WriteUtils extends Logging {
+
+  /** A function that converts the empty string to null for partition values. */
+  case class GpuEmpty2Null(child: Expression) extends GpuUnaryExpression {
+    override def nullable: Boolean = true
+
+    override def doColumnar(input: GpuColumnVector): ColumnVector = {
+      var from: ColumnVector = null
+      var to: ColumnVector = null
+      try {
+        from = ColumnVector.fromStrings("")
+        to = ColumnVector.fromStrings(null)
+        input.getBase.findAndReplaceAll(from, to)
+      } finally {
+        if (from != null) {
+          from.close()
+        }
+        if (to != null) {
+          to.close()
+        }
+      }
+    }
+
+    override def dataType: DataType = StringType
+  }
+
+  def convertGpuEmptyToNull(
+      output: Seq[Attribute],
+      partitionSet: AttributeSet): List[NamedExpression] = {
+    var needConvert = false
+    val projectList: List[NamedExpression] = output.map {
+      case p if partitionSet.contains(p) && p.dataType == StringType && p.nullable =>
+        needConvert = true
+        GpuAlias(GpuEmpty2Null(p), p.name)()
+      case other => other
+    }.toList // Force list to avoid recursive Java serialization of lazy list Seq implementation
+
+    if (needConvert) projectList else Nil
+  }
+
+  def hasGpuEmptyToNull(expressions: Seq[Expression]): Boolean = {
+    expressions.exists(_.find(_.isInstanceOf[GpuEmpty2Null]).isDefined)
+  }
+}
